feat(components): use same deprecation warning function for link-social and grid | sas | #2155
