+++ packages/components/src/utils/log/warnIfDeprecatedComponentIsUsed.ts
@@ -1,6 +1,5 @@
 import { getTagNameWithoutPrefix } from '../tag-name';
 
-// TODO: add missing unit test
 export const warnIfDeprecatedComponentIsUsed = (host: HTMLElement, message?: string): void => {
   console.warn(
     `[Porsche Design System] Component "${getTagNameWithoutPrefix(
