+++ packages/components/src/utils/validation/index.ts
@@ -2,7 +2,6 @@ export * from './getOnlyChildHTMLElementOrThrow';
 export * from './getOnlyChildOfKindHTMLElementOrThrow';
 export * from './throwIfChildCountIsExceeded';
 export * from './throwIfChildrenAreNotOfKind';
-export * from './warnIfComponentIsDeprecated';
 export * from './throwIfElementHasAttribute';
 export * from './throwIfInvalidLinkUsage';
 export * from './throwIfParentIsNotOfKind';
