--- packages/components/src/utils/validation/warnIfComponentIsDeprecated.ts
@@ -1,5 +1,0 @@
-import { getTagName } from '../../utils';
-
-export const warnIfComponentIsDeprecated = (host: HTMLElement, message: string): void => {
-  console.warn(`${getTagName(host)}: ${message}`);
-};
